backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1
google-generativeai==0.7.0
openai==1.35.0
anthropic==0.29.0
httpx==0.27.0
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tinydb==4.8.0
pydantic==2.7.4
python-dotenv==1.0.1
Pillow==10.4.0
